refactor: 파일 저장 기능 global로 분리
