+++ src/main/java/com/ray/pominowner/global/vo/FileInfo.java
@@ -1,0 +1,5 @@
+package com.ray.pominowner.global.vo;
+
+public record FileInfo(String originalFileName, String createdFileName) {
+
+}
