+++ src/main/java/com/ray/pominowner/store/service/StoreImageService.java
@@ -1,6 +1,8 @@
 package com.ray.pominowner.store.service;
 
 import com.ray.pominowner.global.config.ImagePathProvider;
+import com.ray.pominowner.global.vo.FileInfo;
+import com.ray.pominowner.global.vo.FileSaver;
 import com.ray.pominowner.store.domain.Store;
 import com.ray.pominowner.store.domain.StoreImage;
 import com.ray.pominowner.store.repository.StoreImageRepository;
@@ -9,20 +11,17 @@ import org.springframework.stereotype.Component;
 import org.springframework.util.Assert;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.List;
-import java.util.UUID;
 
 @Component
 @RequiredArgsConstructor
 public class StoreImageService {
 
-    private static final String DOT = ".";
-
     private final ImagePathProvider imagePathProvider;
 
     private final StoreImageRepository storeImageRepository;
+
+    private final FileSaver fileSaver;
 
     public void saveImages(List<MultipartFile> images, Store store) {
         Assert.noNullElements(images, "올바르지 못한 파일입니다.");
@@ -31,11 +30,9 @@ public class StoreImageService {
     }
 
     private void saveEachFile(Store store, MultipartFile image, String rootPath) {
-        String originalFilename = image.getOriginalFilename();
-        validateFileName(originalFilename);
-
-        String createdFileName = createFileName(originalFilename);
-        saveImageToPath(image, rootPath + createdFileName);
+        FileInfo fileInfo = fileSaver.retrieveFileInfoAfterSavingFile(image, rootPath);
+        final String createdFileName = fileInfo.createdFileName();
+        final String originalFilename = fileInfo.originalFileName();
 
         storeImageRepository.save(StoreImage.builder()
                 .path(rootPath + createdFileName)
@@ -43,34 +40,6 @@ public class StoreImageService {
                 .fileName(createdFileName)
                 .store(store)
                 .build());
-    }
-
-    private void validateFileName(String originalFilename) {
-        Assert.notNull(originalFilename, "올바르지 못한 파일입니다.");
-
-        if (!originalFilename.contains(DOT)) {
-            throw new IllegalArgumentException("올바르지 못한 파일입니다.");
-        }
-    }
-
-    private void saveImageToPath(MultipartFile image, String path) {
-        try {
-            image.transferTo(new File(path));
-        } catch (IOException e) {
-            throw new RuntimeException("파일 저장에 실패했습니다.",e);
-        }
-    }
-
-    private String createFileName(String originalFilename) {
-        String fileExtension = extractFileExtension(originalFilename);
-        return UUID.randomUUID()
-                .toString()
-                .concat(fileExtension);
-    }
-
-    private String extractFileExtension(String originalFilename) {
-        int dotIndex = originalFilename.lastIndexOf(DOT);
-        return originalFilename.substring(dotIndex);
     }
 
 }
